Split related code into modules
